+++ src/main/java/core/model/match/Matchdetails.java
@@ -6,8 +6,11 @@
  */
 package core.model.match;
 
+import core.db.DBManager;
 import core.model.HOVerwaltung;
+import core.net.OnlineWorker;
 import core.util.HOLogger;
+import module.teamAnalyzer.vo.Match;
 
 import java.sql.Timestamp;
 import java.util.ArrayList;
@@ -17,6 +20,7 @@ import java.util.Vector;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import static core.util.StringUtils.getResultString;
 
 
 public class Matchdetails implements core.model.match.IMatchDetails {
@@ -27,7 +31,7 @@ public class Matchdetails implements core.model.match.IMatchDetails {
     private String m_sMatchreport = "";
     private Timestamp m_clFetchDatum;
     private Timestamp m_clSpielDatum;
-    private ArrayList<MatchEvent> m_vHighlights = new ArrayList<>();
+    private ArrayList<MatchEvent> m_vHighlights;
     private int m_iArenaID = -1;
     private int m_iGastId = -1;
 
@@ -105,6 +109,129 @@ public class Matchdetails implements core.model.match.IMatchDetails {
 
     public ArrayList<Injury> m_Injuries = new ArrayList<>();
 
+    /*
+    goals of each part [0..4] of the match (see MatchEvent.MatchPartId)
+     */
+    private Integer[] homeGoalsInParts;
+    private Integer[] guestGoalsInParts;
+
+    public String getResultInPart(MatchEvent.MatchPartId part){
+        if ( homeGoalsInParts != null){
+            return getResultString(homeGoalsInParts[part.getValue()], guestGoalsInParts[part.getValue()], "");
+        }
+        return "";
+    }
+
+    public String getResultAfterPart(MatchEvent.MatchPartId part) {
+        if (homeGoalsInParts != null) {
+            int home = 0;
+            int guest = 0;
+            for (var p = MatchEvent.MatchPartId.BEFORE_THE_MATCH_STARTED.getValue(); p <= part.getValue(); p++) {
+                if (homeGoalsInParts[p] != null) {
+                    home += homeGoalsInParts[p];
+                    guest += guestGoalsInParts[p];
+                }
+            }
+            return getResultString(home, guest, "");
+        }
+        if (part == MatchEvent.MatchPartId.SECOND_HALF && this.getLastMinute() < 110 ||
+                part == MatchEvent.MatchPartId.OVERTIME && this.getLastMinute() < 121 ||
+                part == MatchEvent.MatchPartId.PENALTY_CONTEST) {
+            return getResult();
+        }
+        return "";
+    }
+
+    public String getResult(){
+        if ( this.getFetchDatum() != null){
+            return getResultString(this.m_iHomeGoals, this.m_iGuestGoals,"");
+        }
+        return getResultString(-1,-1,"");
+    }
+
+    // Return match result extension information as abbreviation string
+    // e. g. a.p. means after penalties. a.e. after extension
+    public String getResultExtensionAbbreviation() {
+        return getResultExtensionAbbreviation(getLastMinuteFromPartInfo());
+    }
+
+    private int getLastMinuteFromPartInfo(){
+        if ( homeGoalsInParts != null){
+            if ( homeGoalsInParts[4] != null ) return 130;
+            if ( homeGoalsInParts[3] != null ) return 120;
+            return 90;
+        }
+        return getLastMinute();
+    }
+
+    static String afterPenalties = HOVerwaltung.instance().getLanguageString("ls.match.after.penalties.abbreviation");
+    static String afterExtension = HOVerwaltung.instance().getLanguageString("ls.match.after.extension.abbreviation");
+    public static  String getResultExtensionAbbreviation(Integer duration){
+        if ( duration != null) {
+            if (duration.intValue() > 120)
+                return afterPenalties;
+            if (duration.intValue() > 110)
+                return afterExtension;
+        }
+        return "";
+    }
+
+    // results like : 2 - 4 a.p.
+    public String getResultEx(){
+        return getResultString(this.m_iHomeGoals, this.m_iGuestGoals, getResultExtensionAbbreviation());
+    }
+
+    // results like : 1 - 1 (2 - 4 a.p.)
+    public String getResultLong(){
+        String res = getResultAfterPart(MatchEvent.MatchPartId.SECOND_HALF);
+        if ( getLastMinuteFromPartInfo() > 110){
+            res += " (" + getResultEx() + ")";
+        }
+        return res;
+    }
+
+    /*
+    Reloading match details is real expensive, so there is a limit of maximum reloads per session
+    Reasons to reload match details:
+    - match details do not contain highlights
+    - highlights do not contain match part (column is introduced with #561 in HO V4.0)
+     */
+    private static int maxMatchdetailsReloadsPerSession = 10;
+
+    // Enable explicit download of match details (e. g. invoked by user clicks)
+    public static Matchdetails getMatchdetails(int matchId, MatchType type, boolean force)
+    {
+        if ( force && maxMatchdetailsReloadsPerSession==0){
+            maxMatchdetailsReloadsPerSession=1;
+        }
+        return getMatchdetails(matchId, type);
+    }
+
+    // Matchdetails factory, checks if database object needs an update from chpp (OnlineWorker)
+    // Number of implicite updates per session is limited to maxMatchdetailsReloadsPerSession
+    public static Matchdetails getMatchdetails(int matchId, MatchType type){
+        var ret = DBManager.instance().getMatchDetails(matchId);
+        if ( maxMatchdetailsReloadsPerSession>0 && ret != null && ret.getFetchDatum()!=null && type.isOfficial()) {
+            var events = ret.getHighlights();
+            if (events == null || events.size() == 0 || events.get(0).getMatchPartId() == null) {
+                HOLogger.instance().info(Matchdetails.class,
+                        "Reload Matchdetails id: "+matchId
+                        +" type:" + type.getName()
+                        +" events:" + (events==null? "null": events.size())
+                );
+                try {
+                    OnlineWorker.downloadMatchData(matchId, type, true);
+                    ret = DBManager.instance().getMatchDetails((matchId));
+                }
+                catch (Exception ex){
+                    HOLogger.instance().error(Matchdetails.class, ex.getMessage());
+                }
+                maxMatchdetailsReloadsPerSession--;
+            }
+        }
+        return ret;
+    }
+
     public void setRatingIndirectSetPiecesAtt(int ratingIndirectSetPiecesAtt) {
         this.ratingIndirectSetPiecesAtt = ratingIndirectSetPiecesAtt;
     }
@@ -119,6 +246,83 @@ public class Matchdetails implements core.model.match.IMatchDetails {
 
     public int getRatingIndirectSetPiecesAtt() {
         return ratingIndirectSetPiecesAtt;
+    }
+
+    public int getLastMinute() {
+        if ( getHighlights().size()>0){
+            return getHighlights().get(getHighlights().size()-1).getMinute();
+        }
+        return -1;
+    }
+
+    public Integer getHomeGoalsInPart(MatchEvent.MatchPartId matchPartId) throws Exception {
+        if ( homeGoalsInParts == null){
+            InitGoalsInParts();
+        }
+        return homeGoalsInParts[matchPartId.getValue()];
+    }
+    public Integer getGuestGoalsInPart(MatchEvent.MatchPartId matchPartId) throws Exception {
+        if ( guestGoalsInParts == null){
+            InitGoalsInParts();
+        }
+        return guestGoalsInParts[matchPartId.getValue()];
+    }
+
+    private void InitGoalsInParts() throws Exception {
+        homeGoalsInParts = new Integer[MatchEvent.MatchPartId.values().length];
+        guestGoalsInParts= new Integer[MatchEvent.MatchPartId.values().length];
+        for ( var event : getHighlights()){
+            int part = 0;
+            var partId = event.getMatchPartId();
+            if (partId != null) part = partId.getValue();
+            if ( homeGoalsInParts[part] == null) {
+                homeGoalsInParts[part] = 0;
+                guestGoalsInParts[part] = 0;
+            }
+            var eventId = event.getMatchEventID().getValue();
+            if ( eventId>99 && eventId<200 ||
+                    eventId>54 && eventId<58
+            ) {
+                // goal
+                if (event.getTeamID() == this.m_iHeimId ||
+                        event.getTeamID() <= 0 && this.m_iHeimId <= 0) { // Verlegenheitstruppe has id < 0 are stored as 0 in event
+                        homeGoalsInParts[part]++;
+
+                } else {
+
+                    guestGoalsInParts[part]++;
+                }
+            }
+            else if ( eventId == MatchEvent.MatchEventID.HOME_TEAM_WALKOVER.getValue()){
+                guestGoalsInParts[0] = 5;
+            }
+            else if ( eventId == MatchEvent.MatchEventID.AWAY_TEAM_WALKOVER.getValue()){
+                homeGoalsInParts[0] = 5;
+            }
+        }
+        /*/ check
+        int sumHomeGoals=0;
+        for ( var i : homeGoalsInParts){
+            sumHomeGoals += i;
+        }
+        if ( sumHomeGoals != this.m_iHomeGoals)
+            throw new Exception("error in InitGoalsInParts");
+        int sumGuestGoals=0;
+        for ( var i : guestGoalsInParts){
+            sumGuestGoals += i;
+        }
+        if ( sumGuestGoals != this.m_iGuestGoals)
+            throw new Exception("error in InitGoalsInParts");
+
+         */
+    }
+
+    public void setHomeGoalsInPart(Integer[] homeGoalsInPart) {
+        this.homeGoalsInParts = homeGoalsInPart;
+    }
+
+    public void setGuestGoalsInPart(Integer[] guestGoalsInPart) {
+        this.guestGoalsInParts = guestGoalsInPart;
     }
 
     public enum eInjuryType {
@@ -734,6 +938,16 @@ public class Matchdetails implements core.model.match.IMatchDetails {
      * @return Value of property m_vHighlights.
      */
     public final ArrayList<MatchEvent> getHighlights() {
+        if ( m_vHighlights == null || m_vHighlights.size() == 0){
+            m_vHighlights = DBManager.instance().getMatchHighlights(this.getMatchID());
+            /* recursiv call in downloadMatchData
+            if ( m_vHighlights.size()==0){
+
+                OnlineWorker.downloadMatchData(this.m_iMatchID, this.m_MatchTyp, true);
+                m_vHighlights = DBManager.instance().getMatchHighlights((this.m_iMatchID));
+            }
+        */
+        }
         return m_vHighlights;
     }