+++ src/main/java/core/net/OnlineWorker.java
@@ -136,8 +136,6 @@ public class OnlineWorker {
 									.setAngezeigteAufstellung(new AufstellungCBItem(
 											getLangString("AktuelleAufstellung"), homodel
 													.getLineup()));
-							homf.getAufstellungsPanel().getAufstellungsPositionsPanel()
-									.exportOldLineup("Actual");
 						}
 						// Info
 						setInfoMsg(getLangString("HRFErfolg"));
@@ -371,6 +369,7 @@ public class OnlineWorker {
 								"Error downloading match. Details is null: " + matchid);
 						return false;
 					}
+					info.setDuration(details.getLastMinute());
 					info.setGastTore(details.getGuestGoals());
 					info.setHeimTore(details.getHomeGoals());
 					info.setGastID(lineup.getGastId());
@@ -545,7 +544,6 @@ public class OnlineWorker {
 				waitDialog.setValue(80);
 
 				matches = FilterUserSelection(matches);
-
 				DBManager.instance().storeMatchKurzInfos(
 						matches.toArray(new MatchKurzInfo[matches.size()]));